+++ sndobj/AdSyn.h
@@ -31,8 +31,8 @@ class AdSyn : public ReSyn {
 
   AdSyn();
   AdSyn(SinAnal* input, int maxtracks, Table* table,
-	    float pitch = 1.f, float scale=1.f,
-	    int vecsize=DEF_VECSIZE, float sr=DEF_SR);
+	    double pitch = 1.f, double scale=1.f,
+	    int vecsize=DEF_VECSIZE, double sr=DEF_SR);
   ~AdSyn();
   short DoProcess();
 };
