+++ sndobj/PVA.h
@@ -36,22 +36,22 @@ class PVA : public FFT {
  protected:
 
   int  m_rotcount; // rotation counter 
-  float m_factor;  // conversion factor
-  float* m_phases;
+  double m_factor;  // conversion factor
+  double* m_phases;
 
  private:
 
-  void inline pvanalysis(float* signal); 
+  void inline pvanalysis(double* signal); 
 
  public:
 
   PVA();
-  PVA(Table* window, SndObj* input, float scale=1.f,
-      int fftsize=DEF_FFTSIZE, int hopsize=DEF_VECSIZE, float sr=DEF_SR);
+  PVA(Table* window, SndObj* input, double scale=1.f,
+      int fftsize=DEF_FFTSIZE, int hopsize=DEF_VECSIZE, double sr=DEF_SR);
 
   ~PVA();
-  float Outphases(int pos){ return m_phases[pos]; } // reads phase output.
-  int Set(char* mess, float value);
+  double Outphases(int pos){ return m_phases[pos]; } // reads phase output.
+  int Set(char* mess, double value);
   void SetFFTSize(int fftsize);
   void SetHopSize(int hopsize);
   short DoProcess();
