+++ sndobj/ReSyn.h
@@ -29,18 +29,18 @@ class ReSyn : public SinSyn {
 
  protected:
 
-  float m_pitch;
-  float m_tscal;
+  double m_pitch;
+  double m_tscal;
 
  public:
 
   ReSyn();
-  ReSyn(SinAnal* input, int maxtracks, Table* table, float pitch=1.f, 
-	float scale=1.f, float tscal=1.f, int vecsize=DEF_VECSIZE, 
-	float sr=DEF_SR);
-  void SetPitch(float pitch){ m_pitch = pitch; }
-  void SetTimeScale(float scale) { m_tscal = scale; }
-  int Set(char* mess, float value);
+  ReSyn(SinAnal* input, int maxtracks, Table* table, double pitch=1.f, 
+	double scale=1.f, double tscal=1.f, int vecsize=DEF_VECSIZE, 
+	double sr=DEF_SR);
+  void SetPitch(double pitch){ m_pitch = pitch; }
+  void SetTimeScale(double scale) { m_tscal = scale; }
+  int Set(char* mess, double value);
   ~ReSyn();
   short DoProcess();
